# front/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Le Renard Étudiant — API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/comptes/moi": {
      "get": {
        "tags": [
          "comptes"
        ],
        "summary": "Moi",
        "operationId": "moi_api_comptes_moi_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Moi"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "comptes"
        ],
        "summary": "Modifier Profil",
        "operationId": "modifier_profil_api_comptes_moi_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificationProfil"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Moi"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comptes/moi/suppression": {
      "post": {
        "tags": [
          "mes données"
        ],
        "summary": "Supprimer Mon Compte",
        "description": "Supprime le compte pour de bon. Le mot de passe est redemandé.",
        "operationId": "supprimer_mon_compte_api_comptes_moi_suppression_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandeSuppression"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/relecture/{revision_id}": {
      "get": {
        "tags": [
          "relecture"
        ],
        "summary": "Lire Revision",
        "operationId": "lire_revision_api_relecture__revision_id__get",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionARelire"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relecture/decisions": {
      "post": {
        "tags": [
          "relecture"
        ],
        "summary": "Decider En Groupe",
        "description": "Applique la même décision à plusieurs versions. C'est tout ou rien.",
        "operationId": "decider_en_groupe_api_relecture_decisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandeDecisionGroupee"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultatDecisionGroupee"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/relecture/{revision_id}/decision": {
      "post": {
        "tags": [
          "relecture"
        ],
        "summary": "Decider",
        "operationId": "decider_api_relecture__revision_id__decision_post",
        "parameters": [
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandeDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionARelire"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Decision": {
        "type": "string",
        "enum": [
          "publier",
          "corriger",
          "refuser"
        ],
        "title": "Decision",
        "description": "Les trois décisions d'un relecteur (ADR 0010)."
      },
      "DecisionLue": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "relecteur": {
            "type": "string",
            "title": "Relecteur"
          },
          "commentaire": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "cree_le": {
            "type": "string",
            "format": "date-time",
            "title": "Cree Le"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "relecteur",
          "commentaire",
          "cree_le"
        ],
        "title": "DecisionLue"
      },
      "DemandeDecision": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "commentaire": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "DemandeDecision"
      },
      "DemandeDecisionGroupee": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "commentaire": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Commentaire"
          },
          "revision_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Revision Ids"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "revision_ids"
        ],
        "title": "DemandeDecisionGroupee",
        "description": "La même décision pour plusieurs versions, par exemple toutes celles d'un import."
      },
...
      "DemandeSuppression": {
        "properties": {
          "mot_de_passe": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Mot De Passe"
          }
        },
        "type": "object",
        "required": [
          "mot_de_passe"
        ],
        "title": "DemandeSuppression"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModificationProfil": {
        "properties": {
          "pseudo": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "pattern": "^[\\w.-]+$",
            "title": "Pseudo"
          }
        },
        "type": "object",
        "required": [
          "pseudo"
        ],
        "title": "ModificationProfil"
      },
...
      "Moi": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "pseudo": {
            "type": "string",
            "title": "Pseudo"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "licence_acceptee": {
            "type": "boolean",
            "title": "Licence Acceptee"
          },
          "jeton_csrf": {
            "type": "string",
            "title": "Jeton Csrf"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "pseudo",
          "role",
          "licence_acceptee",
          "jeton_csrf"
        ],
        "title": "Moi",
        "description": "L'utilisateur connecté. Le jeton CSRF est à renvoyer dans l'en-tête X-CSRF-Token."
      },
      "Niveau": {
        "type": "string",
        "enum": [
          "debutant",
          "intermediaire"
        ],
        "title": "Niveau"
      },
...
      "ResultatDecisionGroupee": {
        "properties": {
          "revision_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Revision Ids"
          }
        },
        "type": "object",
        "required": [
          "revision_ids"
        ],
        "title": "ResultatDecisionGroupee"
      },
...
      "RevisionARelire": {
        "properties": {
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "lecon_slug": {
            "type": "string",
            "title": "Lecon Slug"
          },
          "numero": {
            "type": "integer",
            "title": "Numero"
          },
          "statut": {
            "$ref": "#/components/schemas/StatutRevision"
          },
          "titre": {
            "type": "string",
            "title": "Titre"
          },
          "resume": {
            "type": "string",
            "title": "Resume"
          },
          "objectifs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objectifs"
          },
          "duree_minutes": {
            "type": "integer",
            "title": "Duree Minutes"
          },
          "contenu": {
            "type": "string",
            "title": "Contenu"
          },
          "theme": {
            "$ref": "#/components/schemas/ThemeLecture"
          },
          "niveau": {
            "$ref": "#/components/schemas/Niveau"
          },
          "auteur": {
            "type": "string",
            "title": "Auteur"
          },
          "assiste_par_ia": {
            "type": "boolean",
            "title": "Assiste Par Ia"
          },
          "fichier_importe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fichier Importe"
          },
          "a_relire": {
            "type": "boolean",
            "title": "A Relire"
          },
          "peut_publier": {
            "type": "boolean",
            "title": "Peut Publier"
          },
          "historique": {
            "items": {
              "$ref": "#/components/schemas/VersionHistorique"
            },
            "type": "array",
            "title": "Historique"
          }
        },
        "type": "object",
        "required": [
          "revision_id",
          "lecon_slug",
          "numero",
          "statut",
          "titre",
          "resume",
          "objectifs",
          "duree_minutes",
          "contenu",
          "theme",
          "niveau",
          "auteur",
          "assiste_par_ia",
          "fichier_importe",
          "a_relire",
          "peut_publier",
          "historique"
        ],
        "title": "RevisionARelire"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "apprenant",
          "contributeur",
          "relecteur",
          "admin"
        ],
        "title": "Role",
        "description": "Rôles du cadrage, du moins au plus élevé. Chaque rôle a les droits des précédents."
      },
...
      "StatutRevision": {
        "type": "string",
        "enum": [
          "brouillon",
          "en_relecture",
          "a_corriger",
          "publiee",
          "refusee"
        ],
        "title": "StatutRevision",
        "description": "Statuts du circuit de relecture (ADR 0010)."
      },
...
      "ThemeLecture": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "nom": {
            "type": "string",
            "title": "Nom"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "nom"
        ],
        "title": "ThemeLecture"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VersionHistorique": {
        "properties": {
          "revision_id": {
            "type": "string",
            "format": "uuid",
            "title": "Revision Id"
          },
          "numero": {
            "type": "integer",
            "title": "Numero"
          },
          "statut": {
            "$ref": "#/components/schemas/StatutRevision"
          },
          "auteur": {
            "type": "string",
            "title": "Auteur"
          },
          "cree_le": {
            "type": "string",
            "format": "date-time",
            "title": "Cree Le"
          },
          "publiee_le": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publiee Le"
          },
          "en_ligne": {
            "type": "boolean",
            "title": "En Ligne"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionLue"
            },
            "type": "array",
            "title": "Decisions"
          }
        },
        "type": "object",
        "required": [
          "revision_id",
          "numero",
          "statut",
          "auteur",
          "cree_le",
          "publiee_le",
          "en_ligne",
          "decisions"
        ],
        "title": "VersionHistorique"
      }
    }
  }
}